fix(app): split useAsyncJob.spec.ts under 600-line ceiling (#535 S5b)
useAsyncJob.spec.ts had grown to 609 lines (over the repo's 600-line ceiling
applied to every touched file, tests included). Mirror the useResource split:
extract the cohesive S5b request-ownership / poll-generation describe block into
a sibling useAsyncJob.ownership.spec.ts with its own imports + fake-timer
beforeEach/afterEach harness. No test deleted or weakened — total useAsyncJob
test count preserved at 11 (7 original + 4 ownership).

Line counts: useAsyncJob.spec.ts 432, useAsyncJob.ownership.spec.ts 212.
Gates: code-quality-audit clean, type-check:strict 0 errors, vitest 1995 pass /
2 todo, 0 failures (test files 264 -> 265 from the split; test count unchanged).
